requirements.txt
Flask
flask_sqlalchemy
sqlalchemy_utils
sqlalchemy
psycopg2-binary
passlib
itsdangerous
PyJWT
flask_httpauth